move guts into DefaultBackend
